Edit agent feature
